requirements.txt
SQLAlchemy==0.9.2
Twisted==13.2.0
autobahn==0.8.1
psycopg2==2.5.2
pyOpenSSL==0.13.1
zope.interface==4.1.0
sqlamp==0.6
Flask==0.10.1
Flask-Admin==1.0.7
Flask-Login==0.2.10
Flask-SQLAlchemy==1.0
Babel==1.3
Flask-Babel==0.9
Flask-BabelEx==0.9.1
WTForms==1.0.5
pycrypto==2.6.1
pytest==2.5.2
pytest-twisted==1.5